+++ src/main.jsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client'
 import { BrowserRouter as Router, Routes, Route, Navigate, Outlet } from 'react-router-dom';
 import './assets/styles.css'
 import Layout from '@/starter-components/Layout.jsx';
-import MainView from './components/MainView.jsx'
+import HomeView from './components/HomeView.jsx'
 import OtherView from './components/OtherView.jsx'
 import LandingView from '@/starter-components/LandingView.jsx'
 import TextView from '@/starter-components/TextView.jsx'
@@ -41,8 +41,8 @@ createRoot(document.getElementById('root')).render(
         <Routes>
           <Route element={<Layout />}>
             <Route path="/app" element={<ProtectedRoute />}>
-              <Route index element={<Navigate to="main" replace />} />
-              <Route path="main" element={<MainView />} />
+              <Route index element={<Navigate to="home" replace />} />
+              <Route path="home" element={<HomeView />} />
               <Route path="other" element={<OtherView />} />
               <Route path="settings" element={<SettingsView />} />
             </Route>