+++ changelog.md
@@ -68,6 +68,7 @@ fixed icons in tabbar
 added DynamicIcon from lucide-icons
 0.0.5
 added brand header to sidebar
+fixed full page refresh